# src/fleet/ui/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
    "/api/tasks": {
      "get": {
        "summary": "List Tasks",
        "description": "List task summaries, active first then recently-closed (FR-07).",
        "operationId": "list_tasks_api_tasks_get",
        "parameters": [
          {
            "name": "closed_limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 300,
              "title": "Closed Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Task",
        "description": "Create a task via the queue; 201 with the new id.",
        "operationId": "create_task_api_tasks_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateTaskResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/tasks/{task_id}": {
      "get": {
        "summary": "Get Task",
        "description": "One task summary, overlaid with beads status/priority/depends_on.",
        "operationId": "get_task_api_tasks__task_id__get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Task",
        "description": "Delete a task via the queue.",
        "operationId": "delete_task_api_tasks__task_id__delete",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/tasks/{task_id}/artifacts/state": {
      "get": {
        "summary": "Get Artifact State",
        "description": "STATE.md, or the legacy view for old task dirs without one.",
        "operationId": "get_artifact_state_api_tasks__task_id__artifacts_state_get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/{task_id}/artifacts/result": {
      "get": {
        "summary": "Get Artifact Result",
        "description": "Live RESULT.json, else the latest attempt snapshot, else legacy.",
        "operationId": "get_artifact_result_api_tasks__task_id__artifacts_result_get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/{task_id}/artifacts/outputs": {
      "get": {
        "summary": "Get Artifact Outputs",
        "description": "Deliverables under tasks/<id>/outputs/.",
        "operationId": "get_artifact_outputs_api_tasks__task_id__artifacts_outputs_get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutputsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/{task_id}/artifacts/research": {
      "get": {
        "summary": "Get Artifact Research",
        "description": "Job worker's RESEARCH.md (see workers/job.py).",
        "operationId": "get_artifact_research_api_tasks__task_id__artifacts_research_get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/{task_id}/artifacts/design": {
      "get": {
        "summary": "Get Artifact Design",
        "description": "Job worker's DESIGN.md (see workers/job.py).",
        "operationId": "get_artifact_design_api_tasks__task_id__artifacts_design_get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/{task_id}/diff": {
      "get": {
        "summary": "Get Task Diff",
        "description": "git diff of the task's cwd (empty when not a git repo).",
        "operationId": "get_task_diff_api_tasks__task_id__diff_get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiffResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ArtifactResponse": {
        "properties": {
          "content": {
            "type": "string",
            "title": "Content"
          },
          "mtime": {
            "type": "number",
            "title": "Mtime"
          },
          "path": {
            "type": "string",
            "title": "Path"
          }
        },
        "type": "object",
        "required": [
          "content",
          "mtime",
          "path"
        ],
        "title": "ArtifactResponse",
        "description": "File content envelope for the artifact routes."
      },
...
      "CreateTaskResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "CreateTaskResponse",
        "description": "New task id for POST /api/tasks."
      },
      "DiffResponse": {
        "properties": {
          "diff": {
            "type": "string",
            "title": "Diff"
          }
        },
        "type": "object",
        "required": [
          "diff"
        ],
        "title": "DiffResponse",
        "description": "git diff envelope for GET /api/tasks/{id}/diff."
      },
...
      "FollowupItem": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title",
            "default": ""
          },
          "body": {
            "type": "string",
            "title": "Body",
            "default": ""
          },
          "cwd": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cwd"
          },
          "depends_on": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Depends On"
          }
        },
        "type": "object",
        "title": "FollowupItem",
        "description": "One observer-declared follow-up bead inside a partial result."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobArtifacts": {
        "properties": {
          "research": {
            "type": "boolean",
            "title": "Research"
          },
          "design": {
            "type": "boolean",
            "title": "Design"
          },
          "tasks": {
            "type": "boolean",
            "title": "Tasks"
          },
          "approved": {
            "type": "boolean",
            "title": "Approved"
          }
        },
        "type": "object",
        "required": [
          "research",
          "design",
          "tasks",
          "approved"
        ],
        "title": "JobArtifacts",
        "description": "Presence flags for the job worker's RESEARCH/DESIGN/tasks/APPROVED."
      },
...
      "OkResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          }
        },
        "type": "object",
        "required": [
          "ok"
        ],
        "title": "OkResponse",
        "description": "Generic mutation acknowledgement ({\"ok\": true})."
      },
      "OutputsResponse": {
        "properties": {
          "files": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Files"
          }
        },
        "type": "object",
        "required": [
          "files"
        ],
        "title": "OutputsResponse",
        "description": "Deliverable names for GET /api/tasks/{id}/artifacts/outputs."
      },
...
      "RunStep": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "status": {
            "type": "string",
            "enum": [
              "ok",
              "fail",
              "outcome"
            ],
            "title": "Status"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "name",
          "started_at",
          "ended_at",
          "status",
          "reason"
        ],
        "title": "RunStep",
        "description": "One entry of the current attempt's run.json steps."
      },
...
      "TaskDetail": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "cwd": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cwd"
          },
          "coder": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coder"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "priority": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority"
          },
          "depends_on": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Depends On"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "elapsed_sec": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Elapsed Sec"
          },
          "idle_sec": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idle Sec"
          },
          "events": {
            "type": "integer",
            "title": "Events"
          },
          "context_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context Tokens"
          },
          "context_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context Pct"
          },
          "context_limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context Limit"
          },
          "last_event_kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Event Kind"
          },
          "last_event_detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Event Detail"
          },
          "blocked_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blocked Reason"
          },
          "blocked_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blocked At"
          },
          "ignore_until": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ignore Until"
          },
          "ignored": {
            "type": "boolean",
            "title": "Ignored"
          },
          "rounds": {
            "$ref": "#/components/schemas/TaskRounds"
          },
          "restarts": {
            "type": "integer",
            "title": "Restarts"
          },
          "context_rounds": {
            "type": "integer",
            "title": "Context Rounds"
          },
          "compactions": {
            "type": "integer",
            "title": "Compactions"
          },
          "peak_context_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Peak Context Pct"
          },
          "last_outcome": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Outcome"
          },
          "last_outcome_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Outcome Reason"
          },
          "last_action": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Action"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TaskResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "state_excerpt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State Excerpt"
          },
          "worker": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Worker"
          },
          "job_phase": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Phase"
          },
          "job_artifacts": {
            "$ref": "#/components/schemas/JobArtifacts"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/RunStep"
            },
            "type": "array",
            "title": "Steps"
          },
          "lease": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TaskLease"
              },
              {
                "type": "null"
              }
            ]
          },
          "attempts": {
            "items": {
              "$ref": "#/components/schemas/TaskAttempt"
            },
            "type": "array",
            "title": "Attempts"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "description",
          "status",
          "cwd",
          "coder",
          "model",
          "priority",
          "depends_on",
          "created_at",
          "started_at",
          "ended_at",
          "elapsed_sec",
          "idle_sec",
          "events",
          "context_tokens",
          "context_pct",
          "context_limit",
          "last_event_kind",
          "last_event_detail",
          "blocked_reason",
          "blocked_at",
          "ignore_until",
          "ignored",
          "rounds",
          "restarts",
          "context_rounds",
          "compactions",
          "peak_context_pct",
          "last_outcome",
          "last_outcome_reason",
          "last_action",
          "result",
          "state_excerpt",
          "worker",
          "job_phase",
          "job_artifacts",
          "steps",
          "lease",
          "attempts"
        ],
        "title": "TaskDetail",
        "description": "One task with its attempts timeline (GET /api/tasks/{id})."
      },
...
      "TaskLease": {
        "properties": {
          "heartbeat_at": {
            "type": "string",
            "title": "Heartbeat At"
          },
          "lease_until": {
            "type": "string",
            "title": "Lease Until"
          },
          "alive": {
            "type": "boolean",
            "title": "Alive"
          }
        },
        "type": "object",
        "required": [
          "heartbeat_at",
          "lease_until",
          "alive"
        ],
        "title": "TaskLease",
        "description": "Claim lease from the latest attempt's run.json, null when absent."
      },
      "TaskListResponse": {
        "properties": {
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/TaskDetail"
            },
            "type": "array",
            "title": "Tasks"
          }
        },
        "type": "object",
        "required": [
          "tasks"
        ],
        "title": "TaskListResponse",
        "description": "Envelope for GET /api/tasks."
      },
      "TaskResult": {
        "properties": {
          "schema": {
            "type": "integer",
            "title": "Schema"
          },
          "status": {
            "type": "string",
            "enum": [
              "done",
              "partial",
              "blocked"
            ],
            "title": "Status"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "commits": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Commits"
          },
          "tests": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tests"
          },
          "open_questions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Open Questions"
          },
          "next_step": {
            "type": "string",
            "title": "Next Step"
          },
          "blocked_reason": {
            "type": "string",
            "title": "Blocked Reason"
          },
          "followups": {
            "items": {
              "$ref": "#/components/schemas/FollowupItem"
            },
            "type": "array",
            "title": "Followups"
          }
        },
        "type": "object",
        "required": [
          "schema",
          "status",
          "summary",
          "commits",
          "tests",
          "open_questions",
          "next_step",
          "blocked_reason"
        ],
        "title": "TaskResult",
        "description": "The worker's declared outcome (live RESULT.json or latest snapshot)."
      },
      "TaskRounds": {
        "properties": {
          "failure": {
            "type": "integer",
            "title": "Failure"
          },
          "stall": {
            "type": "integer",
            "title": "Stall"
          },
          "context": {
            "type": "integer",
            "title": "Context"
          },
          "partial": {
            "type": "integer",
            "title": "Partial"
          },
          "noclose": {
            "type": "integer",
            "title": "Noclose"
          }
        },
        "type": "object",
        "required": [
          "failure",
          "stall",
          "context",
          "partial",
          "noclose"
        ],
        "title": "TaskRounds",
        "description": "Trailing-streak retry rounds, from attempts.jsonl."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}